Remove PlayerSession class and add RemoteConnection class for managing remote device connections
