--- Assets/Scripts/Core/Session/PlayerSession.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 278271b48ad9623a2a48af933d5f4ae1
